fix: use local magento to test
